extract functions, remove dateCheck and replace with findDate
